[Node] Re-arrange QnAMaker samples and README
- Re-arrange directories in preparation for adding LUIS Action Binding
- Format sample code and converted tab to spaces for consistency with the lib's source code
