+++ Node/samples/QnAMaker/QnAMakerBotWithActiveLearning/app.js
@@ -1,7 +1,6 @@
-
 var restify = require('restify');
 var builder = require('botbuilder');
-var cognitiveservices = require('../../lib/botbuilder-cognitiveservices');
+var cognitiveservices = require('../../../lib/botbuilder-cognitiveservices');
 
 //=========================================================
 // Bot Setup
@@ -10,13 +9,13 @@ var cognitiveservices = require('../../lib/botbuilder-cognitiveservices');
 // Setup Restify Server
 var server = restify.createServer();
 server.listen(process.env.port || process.env.PORT || 3978, function () {
-   console.log('%s listening to %s', server.name, server.url); 
+    console.log('%s listening to %s', server.name, server.url); 
 });
   
 // Create chat bot
 var connector = new builder.ChatConnector({
- appId: process.env.MICROSOFT_APP_ID,
- appPassword: process.env.MICROSOFT_APP_PASSWORD
+    appId: process.env.MICROSOFT_APP_ID,
+    appPassword: process.env.MICROSOFT_APP_PASSWORD
 });
 var bot = new builder.UniversalBot(connector);
 server.post('/api/messages', connector.listen());
@@ -26,18 +25,18 @@ server.post('/api/messages', connector.listen());
 //=========================================================
 
 var recognizer = new cognitiveservices.QnAMakerRecognizer({
-	knowledgeBaseId: 'set your kbid here', 
-	subscriptionKey: 'set your subscription key here',
-	top: 3});
+    knowledgeBaseId: '67f6513c-8abc-4563-944e-07f5da2ff187', 
+    subscriptionKey: '6cbf13a0503c472c805ab0354317f75e',
+    top: 3});
 
 var qnaMakerTools = new cognitiveservices.QnAMakerTools();
 bot.library(qnaMakerTools.createLibrary());
 
 var basicQnAMakerDialog = new cognitiveservices.QnAMakerDialog({
-	recognizers: [recognizer],
-	defaultMessage: 'No match! Try changing the query terms!',
-	qnaThreshold: 0.3,
-	feedbackLib: qnaMakerTools
+    recognizers: [recognizer],
+    defaultMessage: 'No match! Try changing the query terms!',
+    qnaThreshold: 0.3,
+    feedbackLib: qnaMakerTools
 });
 
 bot.dialog('/', basicQnAMakerDialog);
