+++ Node/samples/QnAMaker/QnAMakerWithCustomTools/app.js
@@ -1,7 +1,7 @@
-
 var restify = require('restify');
 var builder = require('botbuilder');
-var cognitiveservices = require('../../lib/botbuilder-cognitiveservices');
+var cognitiveservices = require('../../../lib/botbuilder-cognitiveservices');
+var customQnAMakerTools = require('./CustomQnAMakerTools');
 
 //=========================================================
 // Bot Setup
@@ -10,13 +10,13 @@ var cognitiveservices = require('../../lib/botbuilder-cognitiveservices');
 // Setup Restify Server
 var server = restify.createServer();
 server.listen(process.env.port || process.env.PORT || 3978, function () {
-   console.log('%s listening to %s', server.name, server.url); 
+    console.log('%s listening to %s', server.name, server.url); 
 });
   
 // Create chat bot
 var connector = new builder.ChatConnector({
- appId: process.env.MICROSOFT_APP_ID,
- appPassword: process.env.MICROSOFT_APP_PASSWORD
+    appId: process.env.MICROSOFT_APP_ID,
+    appPassword: process.env.MICROSOFT_APP_PASSWORD
 });
 var bot = new builder.UniversalBot(connector);
 server.post('/api/messages', connector.listen());
@@ -26,13 +26,18 @@ server.post('/api/messages', connector.listen());
 //=========================================================
 
 var recognizer = new cognitiveservices.QnAMakerRecognizer({
-	knowledgeBaseId: 'set your kbid here', 
-	subscriptionKey: 'set your subscription key here'});
-	
+    knowledgeBaseId: 'set your kbid here', 
+    subscriptionKey: 'set your subscription key here',
+    top: 4});
+
+var customQnAMakerTools = new customQnAMakerTools.CustomQnAMakerTools();
+bot.library(customQnAMakerTools.createLibrary());
+    
 var basicQnAMakerDialog = new cognitiveservices.QnAMakerDialog({
-	recognizers: [recognizer],
-	defaultMessage: 'No match! Try changing the query terms!',
-	qnaThreshold: 0.3
+    recognizers: [recognizer],
+    defaultMessage: 'No match! Try changing the query terms!',
+    qnaThreshold: 0.3,
+    feedbackLib: customQnAMakerTools
 });
 
 bot.dialog('/', basicQnAMakerDialog);
