+++ Node/samples/QnAMaker/QnAMakerWithFunctionOverrides/app.js
@@ -1,0 +1,60 @@
+var restify = require('restify');
+var builder = require('botbuilder');
+var cognitiveservices = require('../../../lib/botbuilder-cognitiveservices');
+
+//=========================================================
+// Bot Setup
+//=========================================================
+
+// Setup Restify Server
+var server = restify.createServer();
+server.listen(process.env.port || process.env.PORT || 3978, function () {
+    console.log('%s listening to %s', server.name, server.url); 
+});
+  
+// Create chat bot
+var connector = new builder.ChatConnector({
+    appId: process.env.MICROSOFT_APP_ID,
+    appPassword: process.env.MICROSOFT_APP_PASSWORD
+});
+var bot = new builder.UniversalBot(connector);
+server.post('/api/messages', connector.listen());
+
+//=========================================================
+// Bots Dialogs
+//=========================================================
+
+var recognizer = new cognitiveservices.QnAMakerRecognizer({
+    knowledgeBaseId: 'set your kbid here', 
+    subscriptionKey: 'set your subscription key here',
+    top: 4});
+
+var qnaMakerTools = new cognitiveservices.QnAMakerTools();
+bot.library(qnaMakerTools.createLibrary());
+
+var basicQnAMakerDialog = new cognitiveservices.QnAMakerDialog({
+    recognizers: [recognizer],
+    defaultMessage: 'No match! Try changing the query terms!',
+    qnaThreshold: 0.3,
+    feedbackLib: qnaMakerTools
+});
+
+// Override to also include the knowledgebase question with the answer on confident matches
+basicQnAMakerDialog.respondFromQnAMakerResult = function(session, qnaMakerResult){
+    var result = qnaMakerResult;
+    var response = 'Here is the match from FAQ:  \r\n  Q: ' + result.answers[0].questions[0] + '  \r\n A: ' + result.answers[0].answer;
+    session.send(response);
+}
+
+// Override to log user query and matched Q&A before ending the dialog
+basicQnAMakerDialog.defaultWaitNextMessage = function(session, qnaMakerResult){
+    if(session.privateConversationData.qnaFeedbackUserQuestion != null && qnaMakerResult.answers != null && qnaMakerResult.answers.length > 0 
+        && qnaMakerResult.answers[0].questions != null && qnaMakerResult.answers[0].questions.length > 0 && qnaMakerResult.answers[0].answer != null){
+            console.log('User Query: ' + session.privateConversationData.qnaFeedbackUserQuestion);
+            console.log('KB Question: ' + qnaMakerResult.answers[0].questions[0]);
+            console.log('KB Answer: ' + qnaMakerResult.answers[0].answer);
+        }
+    session.endDialog();
+}
+
+bot.dialog('/', basicQnAMakerDialog);
