+++ Node/samples/QnAMaker/QnAWithLUIS/app.js
@@ -1,6 +1,6 @@
 var restify = require('restify');
 var builder = require('botbuilder');
-var cognitiveservices = require('../../lib/botbuilder-cognitiveservices');
+var cognitiveservices = require('../../../lib/botbuilder-cognitiveservices');
 
 //=========================================================
 // Bot Setup
@@ -9,13 +9,13 @@ var cognitiveservices = require('../../lib/botbuilder-cognitiveservices');
 // Setup Restify Server
 var server = restify.createServer();
 server.listen(process.env.port || process.env.PORT || 3978, function () {
-   console.log('%s listening to %s', server.name, server.url); 
+    console.log('%s listening to %s', server.name, server.url); 
 });
   
 // Create chat bot
 var connector = new builder.ChatConnector({
-appId: process.env.MICROSOFT_APP_ID,
-appPassword: process.env.MICROSOFT_APP_PASSWORD
+    appId: process.env.MICROSOFT_APP_ID,
+    appPassword: process.env.MICROSOFT_APP_PASSWORD
 });
 var bot = new builder.UniversalBot(connector);
 server.post('/api/messages', connector.listen());
@@ -25,8 +25,8 @@ server.post('/api/messages', connector.listen());
 //=========================================================
 
 var qnarecognizer = new cognitiveservices.QnAMakerRecognizer({
-	knowledgeBaseId: 'set your kbid here', 
-	subscriptionKey: 'set your subscription key here',
+    knowledgeBaseId: 'set your kbid here', 
+    subscriptionKey: 'set your subscription key here',
     top: 4});
 
 var model='set your luis model uri';
@@ -52,5 +52,5 @@ intents.matches('qna', [
 intents.onDefault([
     function(session){
         session.send('Sorry!! No match!!');
-	}
+    }
 ]);
